refactor(app): extract package from monolith
