+++ shyne_app/__init__.py
@@ -1,0 +1,10 @@
+from .app import app
+from .extensions import csrf, db, login_manager
+from .config import *
+from .models import *
+from .access import *
+from .auth import *
+from .admin import MODEL_REGISTRY, admin, register_admin_views
+from .cli import *
+
+__all__ = [name for name in globals() if not name.startswith("_")]
