+++ shyne_app/extensions.py
@@ -1,0 +1,25 @@
+from pathlib import Path
+
+from flask import Flask
+from flask_login import LoginManager
+from flask_sqlalchemy import SQLAlchemy
+from flask_wtf.csrf import CSRFProtect
+
+PROJECT_ROOT = Path(__file__).resolve().parent.parent
+app = Flask(
+    __name__,
+    template_folder=str(PROJECT_ROOT / "templates"),
+    static_folder=str(PROJECT_ROOT / "static"),
+)
+db = SQLAlchemy()
+csrf = CSRFProtect()
+login_manager = LoginManager()
+
+
+def init_extensions(flask_app):
+    db.init_app(flask_app)
+    csrf.init_app(flask_app)
+    login_manager.init_app(flask_app)
+    login_manager.login_view = "login"
+    login_manager.login_message = "Please sign in to continue."
+    login_manager.login_message_category = "info"
